I0425: gesamtstatistik je Wettbewerb - backend fertig

+++ statistics/backend/mkbiza-api/src/main/java/de/egladil/web/mkbiza_api/domain/dto/Gruppierungsitem.java
@@ -15,6 +15,6 @@ public class Gruppierungsitem {
 	private String name;
 
 	@JsonProperty
-	private int anzahl;
+	private long anzahl;
 
 }
